multi collide fix

+++ Assets/Scripts/PlayerHealth.cs
@@ -4,8 +4,8 @@ using UnityEngine;
 
 public class PlayerHealth : MonoBehaviour
 {
-    public int maxHealth = 5;
-    public int currentHealth;
+    public float maxHealth = 5;
+    public float currentHealth;
     public float time = 1;
     public float timer = 0;
     public bool hasCollide;
